+++ frontend/src/app/auth/guards/login.guard.ts
@@ -1,29 +1,19 @@
 import { Injectable } from '@angular/core';
-import { CanDeactivate, Router, UrlTree } from '@angular/router';
+import { CanActivate, Router, UrlTree } from '@angular/router';
 import { Observable } from 'rxjs';
-import { RestorePageComponent } from '../components/restore-page/restore-page.component';
 import { AuthService } from '../../../core/services/auth.service';
 import { NotificationService } from '../../../core/services/notification.service';
 
 @Injectable({
   providedIn: 'root'
 })
-export class RestoreGuard implements CanDeactivate<RestorePageComponent> {
+export class LoginGuard implements CanActivate {
   constructor(
     private router: Router,
     private authService: AuthService,
     private notificationService: NotificationService,
   ) {}
-  canDeactivate(
-    component: RestorePageComponent,
-  ):
-    | Observable<boolean | UrlTree>
-    | Promise<boolean | UrlTree>
-    | boolean
-    | UrlTree {
-    if (!component.requestForm.dirty) return true;
-    return window.confirm('Data would not be saved. Continue?');
-  }
+
   canActivate(): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
     if (this.authService.areTokensExist()) {
       this.router.navigate(['/organizations']);
